refactor: 💡 move get legs from chain to instruction entity
